feat: credits discount

+++ src/main/java/org/fffd/l23o6/service/OrderService.java
@@ -7,7 +7,7 @@ import org.fffd.l23o6.pojo.enum_.PaymentType;
 import org.fffd.l23o6.pojo.vo.order.OrderVO;
 
 public interface OrderService {
-    Long createOrder(String username, Long trainId, Long fromStationId, Long toStationId, String seatType, Long seatNumber, int price);
+    Long createOrder(String username, Long trainId, Long fromStationId, Long toStationId, String seatType, Long seatNumber, double price);
     List<OrderVO> listOrders(String username);
     OrderVO getOrder(Long id) throws AlipayApiException;
     void cancelOrder(Long id) throws AlipayApiException;
